move methods from client.py to client_methods.py
